refactor quarot/spinquant code with simplying code.
