+++ auto_round/algorithms/transforms/base.py
@@ -172,7 +172,7 @@ def _ensure_registry_populated() -> None:
 # ---------------------------------------------------------------------------
 
 
-class RotationSerializer(ABC):
+class SerializerMixin(ABC):
     """Mixin interface for rotation method serialization.
 
     Any :class:`BaseRotation` subclass that needs save/load support should
@@ -185,7 +185,7 @@ class RotationSerializer(ABC):
     Usage::
 
         @BaseRotation.register("my_rotation")
-        class MyRotation(BaseRotation, RotationSerializer):
+        class MyRotation(BaseRotation, SerializerMixin):
             def apply_to_model(self, ...): ...
             def inject_buffers_on_layer(self, ...): ...
             ...
